% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_creation.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert an Excel link map to an igraph object}
\usage{
as_igraph(link_map)
}
\arguments{
\item{link_map}{An Excel link map created by `scan_directory`}
}
\description{
Convert an Excel link map to an igraph object
}
