% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_analysis.R
\name{get_connected_workbooks}
\alias{get_connected_workbooks}
\title{Limit the graph to workbooks that are connected (in either direction) to a
particular workbook}
\usage{
get_connected_workbooks(g, path)
}
\arguments{
\item{g}{An igraph object created by  `scan_directory`}

\item{path}{The path of the workbook you wish to examine}
}
\description{
Limit the graph to workbooks that are connected (in either direction) to a
particular workbook
}
\examples{
\dontrun{
g <- scan_directory("C:/my spreadsheets/")
get_connected_workbooks(g, "C:/my spreadsheets/bar.xlsx")
}

}
