% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_creation.R
\name{get_unique_files}
\alias{get_unique_files}
\title{Get the unique file names from an Excel link map}
\usage{
get_unique_files(link_map)
}
\arguments{
\item{link_map}{An Excel link map created by `scan_directory`}
}
\description{
This includes files that are linked to by other spreadsheet but no longer
exist, and files outside the directory you scanned.
}
\examples{
\dontrun{
excel_map <- scan_directory("C:/my spreadsheets/")
get_unique_files(excel_map)
}

}
