% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_analysis.R
\name{has_broken_dependency}
\alias{has_broken_dependency}
\title{Check whether a workbook is eventually linked to a workbook that no longer exists.
This searches through the full chain of links - i.e. even if all a workbook's
links exist, if one of the linked workbook's links is broken, this will return
TRUE}
\usage{
has_broken_dependency(g, path)
}
\arguments{
\item{g}{An igraph object created by `as_igraph` of the results from `scan_directory`}

\item{path}{The path to the workbook you wish to inspect}
}
\description{
Check whether a workbook is eventually linked to a workbook that no longer exists.
This searches through the full chain of links - i.e. even if all a workbook's
links exist, if one of the linked workbook's links is broken, this will return
TRUE
}
