% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_analysis.R
\name{has_broken_link}
\alias{has_broken_link}
\title{Check whether a workbook has a link to a file that no longer exists}
\usage{
has_broken_link(g, path)
}
\arguments{
\item{g}{An igraph object created by `scan_directory`}

\item{path}{The path to the workbook you wish to inspect}
}
\description{
Check whether a workbook has a link to a file that no longer exists
}
