% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_plotting.R
\name{plot_d3}
\alias{plot_d3}
\title{Create an interactive plot of a spreadsheet graph (using D3)}
\usage{
plot_d3(link_map, color_by = "last_modified")
}
\arguments{
\item{link_map}{An Excel link map created by `scan_directory`}

\item{color_by}{How to colour the graphs. Defaults to `last_modified` categories. Can
alternatively use `basename` to group by spreadsheet names (helps find duplicates)
or pass the full path of a spreadsheet to highlight just it}
}
\description{
Create an interactive plot of a spreadsheet graph (using D3)
}
\examples{
\dontrun{
map <- scan_directory("C:/my spreadsheets/")
plot_d3(map)
}

}
