% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_creation.R
\name{scan_directory}
\alias{scan_directory}
\title{Scan a directory for excel files and enumerate the linked spreadsheets in
any files that are found.}
\usage{
scan_directory(dir, startup_folder)
}
\arguments{
\item{dir}{The directory to search}

\item{startup_folder}{The folder listed under "At startup, open all files in:",
in Excel's Options > Advanced}
}
\description{
Returns a list mapping files (names in the list) to their links.
}
\examples{
\dontrun{
scan_directory("C:/my spreadsheets/")
}
}
