% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_creation.R
\name{scan_file}
\alias{scan_file}
\title{Scan a single Excel file and enumerate the linked spreadsheets in it}
\usage{
scan_file(dir, path)
}
\arguments{
\item{path}{The Excel file to search}

\item{startup_folder}{The folder listed under "At startup, open all files in:",
in Excel's Options > Advanced}
}
\description{
Returns a list mapping the files (names in the list) to its links.
}
\examples{
\dontrun{
scan_directory("C:/my spreadsheets/")
}
}
